+++ HumanitarianAssistance.Entities/Models/LedgerModel.cs
@@ -13,6 +13,6 @@ namespace HumanitarianAssistance.ViewModels.Models
         public DateTime fromdate { get; set; }
         public DateTime todate { get; set; }
         public int RecordType { get; set; }
-        public List<int?> accountLists { get; set; }
+        public List<long?> accountLists { get; set; }
     }
 }
